+++ src/gazebo_mavlink_interface.cpp
@@ -804,17 +804,17 @@ void GazeboMavlinkInterface::SendGroundTruth()
 
 void GazeboMavlinkInterface::GpsCallback(GpsPtr& gps_msg, const int& id) {
   SensorData::Gps gps_data;
-  gps_data.time_utc_usec = gps_msg->time_utc_usec();
+  gps_data.time_utc_usec = static_cast<uint64_t>(gps_msg->time_utc_usec());
   gps_data.fix_type = 3;
-  gps_data.latitude_deg = gps_msg->latitude_deg() * 1e7;
-  gps_data.longitude_deg = gps_msg->longitude_deg() * 1e7;
-  gps_data.altitude = gps_msg->altitude() * 1000.0;
-  gps_data.eph = gps_msg->eph() * 100.0;
-  gps_data.epv = gps_msg->epv() * 100.0;
-  gps_data.velocity = gps_msg->velocity() * 100.0;
-  gps_data.velocity_north = gps_msg->velocity_north() * 100.0;
-  gps_data.velocity_east = gps_msg->velocity_east() * 100.0;
-  gps_data.velocity_down = -gps_msg->velocity_up() * 100.0;
+  gps_data.latitude_deg = static_cast<int32_t>(gps_msg->latitude_deg() * 1e7);
+  gps_data.longitude_deg = static_cast<int32_t>(gps_msg->longitude_deg() * 1e7);
+  gps_data.altitude = static_cast<int32_t>(gps_msg->altitude() * 1000.0);
+  gps_data.eph = static_cast<uint16_t>(gps_msg->eph() * 100.0);
+  gps_data.epv = static_cast<uint16_t>(gps_msg->epv() * 100.0);
+  gps_data.velocity = static_cast<uint16_t>(gps_msg->velocity() * 100.0);
+  gps_data.velocity_north = static_cast<int16_t>(gps_msg->velocity_north() * 100.0);
+  gps_data.velocity_east = static_cast<int16_t>(gps_msg->velocity_east() * 100.0);
+  gps_data.velocity_down = static_cast<int16_t>(-gps_msg->velocity_up() * 100.0);
   // MAVLINK_HIL_GPS_T CoG is [0, 360]. math::Angle::Normalize() is [-pi, pi].
   ignition::math::Angle cog(atan2(gps_msg->velocity_east(), gps_msg->velocity_north()));
   cog.Normalize();
